pyproject.toml
[project]
name = "proj"
version = "0.6.00_alpha"
dependencies = [
    "beaapi>=0.1.0",
    "bs4>=0.0.2",
    "curl-cffi==0.10.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.3.1",
    "selenium>=4.34.2",
    "cffi == 1.17.1",
    "certifi>=2025.7.14",
    "yfinance>=0.2.65",
    "lxml>=6.0.0",
    "pyside6>=6.9.2",
    "pyqtgraph>=0.13.7",
    "python-dateutil>=2.9.0",
    "pandas-stubs==2.3.2.250827",
    "datetime>=5.5",
    "logging>=0.4.9.6",
    "typing>=3.10.0.0",
    "numpy>=2.3.1",
    "python-i18n>=0.3.9",
    "PyYAML>=6.0",
]
